pyproject.toml
[project]
name = "process-flapper-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.1",
    "pyyaml>=6.0",
    "rerun-sdk>=0.30.1",
    "scipy>=1.17.1",
]
